PibbyAPI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
packaging==23.2
requests==2.31.0
python-dotenv==1.0.0
slowapi==0.1.8
python-multipart==0.0.6
aiohttp==3.9.1
networkx==3.2.1
matplotlib==3.8.2
fpdf==1.7.2
pandas==2.1.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 